Put CLI in bliss.api
Exclude hydra CLI interfacing code from pytest codecov.
